Organización de las Clases
Se ha creado por los 3 lenguajes una carpeta en donde se trabajara por separado la codificación de las clases, con el fin de tener mas orden acá perrón a la hora de trabajar
